Implement proper Eval() method
The previous approach had a lot of this code in the main Program class.
This _worked_, but meant the code was much harder to use from other
scenarios (embedding, unit tests etc).

The PerlangInterpreter class now has a proper Eval() method, to which
you pass a ScanErrorHandler, a ParseErrorHandler, and a
ResolveErrorHandler. This makes it fit much better with unit tests,
coming up next.

(I realize while writing this that we should probably add a
RuntimeErrorHandler or something, which we can use to propagate runtime
errors to the caller as well.)