(many) Add REPL banner (#6)
This includes the version number as well (taken from the AssemblyInfo
for now). It would be nice to show the git commit that the version is
built for as well, and distinguish between "real" tagged versions and
in-progress commits. That will however be left as an exercise for a
later time.