(many) Use ResolveError, ParseError etc in error handler delegates
Again, this is mostly paving the way for other changes coming. This one
is pretty nice though, since it decouples the data structure from the
delegate in a nice way. If/when we want to add more fields to one of
these errors, we can do so without having to touch any of the existing
usages, as long as the new fields are not being used by them. OOP at its
finest. :)