(many) Moved console app to separate Perlang.ConsoleApp project
It feels unreasonable to place the console application inside the actual
interpreter project. While it is slightly simpler that way, it means
that for e.g. embedding scenarios we are forcing consumers to reference
a project that already includes an application of itself. When we start
adding more dependencies that are specific to the ConsoleApp, this will
make even more sense.