(tests) Add `for`-tests (#15)
This also included some changes to the interpreter to be able to capture
the output from `print` statements. This is really nice, since it makes
adding the full Lox test suite much simpler (which is something that I
will probably do, a small chunk at a time, for the relevant parts).