(interpreter) Fix variable resolution in for loops
The core problem here was that the `Stack` classes in Java and .NET seem
to work differently. One grows upwards, the other grows downwards. This
meant that the logic that this code used broke, since it relied on the
Java-style semantics. (it was written in Java initially, remember?)

I fixed this by replacing it with a proper `List` instead, which is also
much more performant on .NET than peeking in the middle of a stack. The
code is arguably a bit more ugly though, so we could add a few helper
methods/create our own `Stack` implementation at some point if we feel
like cleaning things up a bit.

Fixes #12