(all) Add argv_pop() function
This would rightfully be something much more like `ARGV.pop()` in Ruby;
a method on an array with a well-defined name. But since we currently
lack support for objects and classes, this is a poor man's workaround
for the time being.

It _does_ work though, and has an appropriate number of tests.