(many) Detect bare expressions more cleanly (#34)
When I wrote the original code for this, I hadn't looked at the solution
provided by @munificent in
https://github.com/munificent/craftinginterpreters/blob/master/note/answers/chapter08_statements.md.
Hence, my suggested approach was rather brute an anything but elegant.
This was being manifest in e.g. the `var` tests not returning the
expected exceptions; this was a side effect of the ugly "parse as
expression first, then list of statements"-approach. This is definitely
a step in the right direction IMO.