(all) Support integer literals, not just doubles (#40)
I decided to use `dynamic` to implement this for now. It's the simplest
way to not having to handle all various combinations of integer types
manually.

Doing it like this made it quite easy in the end, it felt impossible
just some day ago. :)

As for how to handle integer literals, I decided to with the same rules
are C# applies for now.

> If the literal has no suffix, its type is the first of the following
> types in which its value can be represented: int, uint, long, ulong.

(quoting https://docs.microsoft.com/en-us/dotnet/csharp/language-reference/builtin-types/integral-numeric-types)

The suffix stuff is not supported at the moment; we provide no way to
explicitly specify neither `ulong`, `uint` or `double`. This will
probably have to fixed at some point, but just supporting various
integer types and `double` is probably a good start. (We could do the
same for the floating point values btw, to coerce them into `float` if
the encountered floating point literal is small enough.)