(many) Improve error handling in Eval (#49)
While working on #43, I realized that the current way we've been handling these have been a bit bad in that the stack traces from the exceptions being generated on certain kinds of errors while scanning, parsing or resolving variables have been less than optimal. This in turn has made debugging a less pleasant experience than it ought to be.

By changing these to not build up a list of (scan|parse|resolve) errors but instead call the error handler directly, the debugging experience is improved significantly.

Some error types which were not proper exceptions have also been changed to inherit from Exception. This was probably not required to get this working, but still feels reasonable and more consistent than the previous behavior.