(ci) Use abbreviated git SHA in file names
It turned out that using the full SHA makes it work less efficiently
when browsing the file list at http://builds.perlang.org/; the file
names get so long that the suffix (e.g. `win-x64.tar.gz`) isn't visible.

The default length of an abbreviated git SHA is 7 characters, but as
projects grow larger, they tend to need more characters to uniquely
identify objects. Perlang has no ambitions (yet) to reach hundreds of
thousands of commits, but let's make it a bit longer than this anyway to
not make it look ridiculously short...