(interpreter) Extract IBindingFactory interface (#68)
Again, this is a preparational commit to make room for #66. I concluded that we had some if/else logic here where it made some sense to try to use the best parts of OOP instead, namely polymorphism.

Will this have a significant impact on making #66 easier and more achievable? Perhaps, perhaps not, but it felt like a reasonable small step that would be better off extracting from my WIP branch instead of trying to do "too much" at one time. It's not like adding class support is by any means a trivial task anyway, so anything that can help that branch be more focused and to-the-point is a win.