(many) Fix negative number support
This was broken in #44; it turned out that there was an invalid
assumption that all values were `double`s which was the root cause here.
Before we had static typing support, all numbers were indeed `double`s
so that's why it was coded like that. This was an inheritance from Lox.

Fixes #45.