(many) Minor cleanup: renamed Expr.Variable to Expr.Identifier
I realized that it seemed incorrect to name this expression type
"variable" now that it can refer to both variables, functions and
classes. I'm not 100% sure about this though, since there is also a
_token_ named `IDENTIFIER` which is used when constructing `Get`
expressions in a WIP PR I'm working on, but time will whether this
change is a good idea or not.