(many) Do not allow redeclaring globals (#87)
This has been something which was allowed historically; we inherited the semantics from Lox in this case and it was rediscovered now when working on #60. However, I've not been very fond of it. Yes, yes, Scheme allows it, Rust allows it, but we are neither Scheme nor Rust. C# disallows it, and I think a solid case for not supporting these semantics can be made.

Let's disallow it for now, we can always revisit this decision later if we discover this to be a great thing after all.