(tests) Make EvalResult a generic class
The previous approach - a container class with basically one property
which contains the relevant errors and three which were always empty
lists - was nice in that the test code was very explicit;
`result.ValidationErrors` etc is quite readable code. However, it felt a
bit odd. I realized today we could probably get rid of this and simplify
things a bit by making the class be type-parameterized instead.