(docs) Fix download commands to be more copy-paste friendly
The previous way this worked was that the `$ ` part in the beginning of
the cURL oneline would be included when selecting the line in the web
browser, which would break actual execution in bash or any other shell.
By marking the `$ ` part a non-selectable, the experience on this web
page was improved greatly.

Credit to @maple3142 in
https://github.com/theme-next/hexo-theme-next/issues/448#issuecomment-431677117
which gave me the idea on how to solve this, big thanks to you!