(interpreter) Change postfix increment and decrement value semantics (#145)
Before this change, the value of a postfix increment or decrement was
the value of the expression _after_ the assignment of the new value,
unlike languages like Java and C# where the value of the expression is
taken _before_ the assignment. We should be really careful with semantic
differences like this, _when they are not rational and easily
explained_. The "principle of least astonishment" is an important one.