(consoleapp) Fix bug with parsing of incomplete assignments (#161)
Certain incomplete statements were incorrectly being considered valid,
as a result of https://github.com/perlang-org/perlang/pull/34 being
merged. The special-casing in `ParseError` needed to be removed after
that was merged, but was incorrectly retained. This commit fixes that.
It also adds tests to ensure that we don't regress similarly in the future.