(tests) Add architecture test project, using ArchUnitNET
ArchUnit is a Java library that we've been using at work for a while.
ArchUnitNET is the .NET counterpart. It allows to you to declare
architecture rules for an application, to allow these rules to be
validated automatically in e.g. CI runs. Break the rules => CI will
fail, and you will be either forced to add a manual exception or fix the
problem before you are allowed to merge.

The first concrete use case for this was to ensure that all `Expr`
classes implement the `ITokenAware` interface, except for a few specific
classes where it does not make sense.