(interpreter) #162 Ensure RuntimeErrors stops script execution (#165)
This one turned out to be rather nasty and non-trivial. The way we
handled exceptions was basically broken. We propagated them to the
`runtimeErrorHandler` and carried on. For all other error types
(scanner, parser, type validation etc) this makes perfect sense, but not
so with runtime errors. They are _intended_ to halt the script from
further execution.

Closes #162