(parser) Mark certain keywords as reserved for future use
I started thinking about this the other day. Shipping version 0.1.0 of
Perlang with a half-baked `class` implementation makes little sense, if
any, at all. It's much better to just make `class` a reserved word for
now, clearly implying that it's likely to come in a future release but
not giving the completely false impression that it's already here (when
it truly isn't).

I also added a bunch of other reserved words for now, inspired by C and
Java/C#. There might be things here that are not present; this is not a
problem, we'll just add them to 0.2.0. Hopefully this list will shrink
as time passes (and we implement some of these into the language
proper). In a distant future when we ship 1.0, I hope the list would be
to close to empty, unless we know with a _very_ high degree of certainty
that we have things we will want to use for something great in a future
release.

Closes #178