Update perlang-install to detect tls1.2 for curl 7.73+
(Original commit message from
https://github.com/rust-lang/rustup/commit/01fd3a19774e97839f6b3c52075a491ea9479abd#diff-a8a8a5affa9439c6617ca7b2cf8cb43e31b9a7706cc600852dd9c557dd8a00d0
below)

This fixes rust-lang/rustup#2603.  In curl 7.73.0, the help format was
changed so as to only show a summary when calling `curl --help`, while
the old menu can be accessed as `curl --help all`.  This change updates
the `check_help_for` function to detect if the command being checked
contains the curl 7.73.0+ language indicating the use of `--help all`.

If this language is present, `check_help_for` will insert the `all`
category after the `--help` flag, allowing the query to work.  If the
language is not present, then it will insert an empty string following
the `--help` flag, maintaining the same behavior with pre-7.73 versions
of curl.