(tests) Add ResolverTest + refactor Eval() to support its work
Previously, tests which wanted to inspect the AST or other internal work
of the interpreter needed to manually create a bunch of data structures
and then call the appropriate method (like `TypeValidator.Validate()`;
see `TypeValidatorTest` for an example of how awkward this is).

While the approach described above _does_ have an advantage in that it
requires a less functional interpreter/compiler to be able to run the
tests, it does come with a significant disadvantage as well: it's
tedious, it's manual work, and it's boring. It's very unlikely that any
significant amount of tests would be added when the process is so painful.

The approach introduced in this PR makes things significantly better. We
now have appropriate tooling in place (the
`ScanParseAndResolveSingleStatement` method) to help in writing tests
like this.

----

While full E2E/integration tests (which execute a full Perlang
interpreter and inspect the result or makes assertions on certain kind
of errors) are useful for many aspects of the life of an
interpreter/compiler, there are particular areas where it also makes
sense to be able to "straightjacket" your code a bit more. These tests
are in other words not intented to replace full integration tests but to
complement them.