(parser) Support integer literals larger than 64-bit
The simple fix for now is to just parse them as `BitInteger`. An
interesting question is if we should support suffixes like L, U, UL etc,
as in C#. I think the simple answer is "we don't know", and we'll wait
with implementing things like that until there's a good-enough
motivation for it.

Closes #205