(tests) Add data-driven tests for comparison operators (#230)
While working on #227, I noted that there are many different permutations of operations (>, <, <=, >= and so forth) and types where we only had test coverage of the int + int scenarios. It's much better to try and (as far as is reasonably possible) cover all supported combinations of types, especially if/when we start handling these manually instead of letting dynamic work it out for us.