(ci) Use build matrix for release publishing (#244)
When it comes to CI, I'm usually a quite big fan of aggressively aiming
for parallelizing as much as possible of the build pipeline. I much
rather have 5 jobs that run for 30-60 seconds each than one job which
runs for 150-300 seconds. It does add a bit of complexity, but I think
the benefits typically outweigh that part.

Now that we are changing from 3 supported architecture (`linux-x64`,
`osx-x64` and `win-x64`) to five (the previous ones + `linux-arm` and
`linux-arm64`), being able to run these builds in parallel does make
sense. It also makes the YAML shorter and sweeter, and it makes it
simpler to add upcoming new architectures in the future. (`osx-arm64` is
the most likely one to be added quite soon.)