(parser) Changed ScanToken to be sorted by ASCII order (#249)
I realized the previous order was rather unpredictable. Many software
engineers have a tendency to prefer implementations where some form of
sorting is applied; I am no exception to this rule.

In this particular case [pun unintentiona], I think it can be motived on
fairly good grounds. It makes it easier for people looking for how a
particular character is handled, and easier if/when we implement a new
token for a previously unhandled character. In other words, it seems to
have more advantages than disadvantages.