(ci) Run tests on all supported platforms (#257)
It is likely that we will soon introduce some functionality which is
OS-dependent (e.g. making it possible to support a subset of the POSIX
system calls); these changes are unlikely to work on Windows. To ensure
a consistent (or at least _predictable_) experience on all platforms,
it's important that we consistently run all tests on all supported
platforms from now on.