(interpreter) Support platform-specific global classes
We will likely never get full POSIX support on Winodws; this is simply a
fact that we'll have to live with. Different languages and runtimes take
different approaches to this:

- Java and .NET try to abstract away the fact that operating systems
  differ, to various degrees. Java sometimes takes this to rather absurd
  lengths, by even making it impossible to do simple things like calling
  `getuid`. "Write once, run anywhere" definitely comes at a price.

- .NET have traditionally done it slightly differently, given that it
  was originally very tied to Windows. Relying heavily on Windows-only
  APIs like GDI+ and DirectX was not a problem in this context. Now that
  .NET is cross-platform, things are indeed a bit different and perhaps
  moving slightly more towards Java (while still providing
  Windows-specifics _for apps that need it_).

Perlang is likely to land more close to the .NET position on this. I
really like the idea of providing a `Posix` module/class, with things
like `getgid`, `getuid`, `getpid` and so forth readily available. When
we support imports, I even imagine being able to do something like
`import Posix.*` or `import static Posix.*`, to be able to call these
methods pretty much exactly the same as in C.

Imagine this for a moment. Something that "feels like C" (in a positive
sense), but still can run in a managed context, even in interpreted
mode, doesn't that sound pretty awesome to you? (Someone will probably
soon start saying "you're reinventing Python, you know that?" to me
soon; I will do my very best to not pay any attention to those voices at
all. :D)