(parser) Fix broken support for certain negative integer literals
This commit changes the logic to do the parsing of numeric literals at
parse-time instead of scan-time, where we have better access to the
actual context of how the literal is being used. This makes it possible
to avoid making expressions like "-100" be `Expr.UnaryPrefix` and
instead be a simple `Expr.Literal` with the value of the expected type,
even for the edge cases described in #302.

Fixes #302