(tests) Improve coverage of binary operator type combinations (#313)
This is done using a new `BinaryOperatorData` class. The purpose of this
class is to have a centralized location where "all" test data for the
binary operators can be placed. This way, when adding a new type you can
go through all the binary operators in a single place and add test cases
for the newly added type, ensuring that it works correctly with them
all.