(tests) Verify expected types in binary operator tests
We previously had some of these types checked, but they were kept in a
separate list from the values, making it more painful to maintain and
much harder to ensure that all values are present in both lists. This
commit refactors this to use _one_ unified list for both the expected
values and the expectd types, with some syntactic sugar to present it to
the tests with "only the expected parameters". This should be a much
more maintainable solution in the long run.