(stdlib) Use better format in `float` and `double` conversion to string (#414)
The main benefit of this new format is that it's "more" predictable, and
we can more easily mimic the exact byte-by-byte output in the C++ implementation.

The downside is that we *may* have problems with roundtripping to string
and back with this format, because of precision loss. Let's live with
this for now; it's (for now) more important that all integration tests
pass in compiled and interpreted mode. .NET Framework seems to have been
using G7 and G15 by default, so it can't be that bad. ;)