(consoleapp) Remove REPL code
As discussed in https://github.com/perlang-org/perlang/issues/406, the
REPL will go away for some time, until we can (at some point)
reimplement it on top of LLVM. At that point, the REPL will be
_dynamically emitting native code_, i.e. still not require a JIT
interpreter of any form. Based on some experiments I've done with LLVM,
this should be doable.

The `-e "<code-to-be-executed>"` will also be removed for now, but will
take that as a separate commit.