(consoleapp) Remove `-e <script>` option (#447)
Similar to the REPL, this currently depends on our `PerlangInterpreter`,
i.e. interpreted mode execution of Perlang code. The ideal would be to
actually be able to remove that class altogether (and at some point
replace it with a Perlang counterpart which interfaces with LLVM
on-the-fly). I don't know if that is feasible or not, but let's at least
take these simple steps first and remove things which obviously depend
on parts of our tooling that we would want to see "go away", so to
speak.