(stdlib) `BigInt::pow`: Detect negative exponents and throw an exception (#450)
This was detected while working on the `EvalHelper` methods to try and
make more of them work in compiled mode too. This would be a noble goal,
but for now feels kind of non-trivial. Because of this, the change
likely does unfortunately not have excessive test coverage at the
moment; we'll just have to take the compiler's word (Clang) for it that
it works. :-)