(compiler) Add `-c` flag for "compile and assemble only" (#455)
This commit is a small step towards #454. By adding this flag, it's
possible to "use Perlang as a transpiler" only. Okay, a _little_ more
than a transpiler, since it also emits ELF object files. :) But still,
it helps us towards a state where we can gradually, incrementally
rewrite parts of the Perlang compiler in Perlang. Oh, the joy! Oh, the
glory!