(compiler) Remove `using namespace` in generated C++ code
This has turned out to be a bad idea. Including this makes it slightly
harder to copy fragments of the generated C++ code into e.g. a test
program, which is something I do from time to time (when e.g. debugging
generated C++ code that doesn't work or doesn't compile).