(compiler) Add partial inline C++ support
This change adds support for writing entire methods in C++. It lets you
declare the prototype(s), using a special preprocessor directive, and
similarly for the actual method implemention.

What we don't support is something like:

cpp {
  // C++ code here inside a Perlang function
}

...and this is also something we don't see a strong need for right now.
Will leave this as-is, and might consider implementing the above at some
point if the need arises.