(compiler) Make compiled mode be the default and drop interpreted mode (#465)
This is the biggest change for a while. Because
https://github.com/perlang-org/perlang/issues/406 is moving along
nicely, we are now ready to:

* Flip the switch, i.e. _make compiled mode the default_ for Perlang.
* Remove the PerlangInterpreter class in its entirety. This may be
  reimplemented in one form or another, once we have the LLVM-emitting
  backend in place, but not as a tree-walking interpreter.

This probably means we'll drop Windows (and perhaps macOS) support for a
while. Please don't despair; this is not intended to be permanent. While
we depend on a specific Clang version for compiling Perlang code, it
simply gets easier to not have to support too many platforms. Once we
have started emitting C++ code from Perlang, in an idempotent way (being
able to disable all timestamping etc in the file header), we could see
how hard it would be to get this Perlang-to-C++-transpiled code
compiling on macOS and Windows too.