(many) Fix ASCIIString+ASCIIString concatenation to return ASCIIString (#474)
The problem with letting it return `String` (as was the case before this
fix) is that the result couldn't be assigned to an `ASCIIString`
variable, but the `TypeResolver` would still believe the result of such
an operation would be a pointer to `ASCIIString`. The end result would
be that we would emit uncompilable C++ code.