(many) Remove char_at() kludge (#493)
This was a workaround to the fact that unwrapping a std::shared_ptr
seemed awkward at the time. However, doing it like this is actually
quite straightforward and feels like a great cleanup. It also means we
don't have to pollute the public API in ASCIIString, which is a great
improvement.