(compiler) Add support for compiling multiple Perlang files
This is really crude and a big hack, but here we go:

- Specifying multiple files on the command line now works, like this:
  perlang -c file1.per file2.per -o some_file.o
- The some_file.o is also taken as the name of the output .cc file. In
  the example above, the transpiled C++ version of the Perlang files
  will be written to some_file.cc. If no -o parameter is used, the first
  name of the input files will be used (file1.per in our example, so the
  output would get written to file1.cc in that case)

The way the files are combined is extremely simplistic for now: the
files are read in the order given on the command line. Their content is
appended into a big string, which is then compiled. For simpler
multi-file programs, this can work but there is no way to
"forward-reference" a method being defined in a "later" file, for
example. We should do an effort at doing a multi-pass compile at some
point, which should help sorting out problems like that.

https://gitlab.perlang.org/perlang/perlang/-/merge_requests/508