(cli) Add first step towards rewriting command line handling in Perlang
...with a bit of C++ help to be able to parse the command line arguments
with `getopt_long` (which is not available on Windows btw, so we need to
take this into account when we make Perlang Windows-compatible again).

This was quite an effort in fact; it was started already in April but
only now reached the point where it's about to be mergable. (I added the
string array support recently, but ended up not having to use it for
now, since we as mentioned use the `getopt_long` stuff for doing the
actual argument parsing for us)

Only a single argument is handled on the Perlang side for now: `-v`.
We'll work our way now through the remaining ones, one by one, now that
the foundation is laid for us to implement this in pure Perlang
goodness. :)

https://gitlab.perlang.org/perlang/perlang/-/merge_requests/506