(cli) Remove hardwired versions in `perlang_cli.cc`
This was unintentionally merged as part of
https://gitlab.perlang.org/perlang/perlang/-/merge_requests/506. The
idea was to keep the version in Git using `<PLACEHOLDER>` strings, which
would then be replaced with proper values when building a snapshot or
release build. I noted this because a snapshot build was displaying this
version, despite the `perlang_cli.cc` file not being rebuilt properly.