(compiler) Remove unused `Perlang.Compiler` project
This was added late last year, when we started working on the
experimental compiler. At that time, I think I was hoping we could put
this in a project/assembly of its own. This turned out to be non-trivial
so I ended up taking the shorter route for now and put it in
`Perlang.Interpreter` for now. Hence, this project has AFAIK never been
used. Let's kill it with 🔥 now.

https://gitlab.perlang.org/perlang/perlang/-/merge_requests/511