(cli) Refactor "autogenerate version strings" code
Previously, this would be embedded in `perlang_cli.cc`; running the
right command would trigger this file to be regenerated, including the
version strings as-is in the file. This made it easy to commit this file
by mistake, as in 08e55f5ef326b69782479605f067a78e434eb07a.

The changed approach means that we generate a special
`perlang_cli_preprocessed.cc` file in the `Makefile` instead, by
replacing certain magic strings in the file with their appropriate
values. This file is never commited into git, and thus we've made it
impossible to commit a "dirty" file by mistake.

https://gitlab.perlang.org/perlang/perlang/-/merge_requests/516