(ci) Disable install tests on macOS
...since macOS builds are broken at the moment. We should probably do
the only sane thing and disable macOS (and Windows) builds for the
moment. We'll for sure try to reintroduce them at some point, but for
now we'll be able to move with greater velocity by keeping it simple and
stupid, and focus on Linux only. Once we are self-hosted and independent
of clang for generating native binaries from Perlang programs, we can
pick this up again.

https://gitlab.perlang.org/perlang/perlang/-/merge_requests/521