(stdlib) Return `unique_ptr<T>` instead of `shared_ptr<T>` when possible
I learned that `unique_ptr<T>` can easily be auto-converted to
`shared_ptr<T>`, but not the other way around. Since we sometimes want
to "consume" the object returned from one of these methods, it makes
more sense to default to `unique_ptr<T>` and let the caller "make it
shared" if required.

Credits to this SO thread, which discusses these conversions:
https://stackoverflow.com/questions/37884728/does-c11-unique-ptr-and-shared-ptr-able-to-convert-to-each-others-type

https://gitlab.perlang.org/perlang/perlang/-/merge_requests/529