(tests) Move Valgrind output to `EvalException`
...instead of printing it to `stderr`. The previous approach was
problematic, since e.g. the `EvalReturningOutputString()` method returns
the output written by the process to the caller. This meant that when
Valgrind mode was enabled, the Valgrind output would be joined together
with the "real" output of the program being tested, messing everything
up completely.

(This is a fixup of the functionality introduced in
https://gitlab.perlang.org/perlang/perlang/-/merge_requests/530)

https://gitlab.perlang.org/perlang/perlang/-/merge_requests/532