(stdlib) De-duplicate string concatenation operator implementations
In the previous implementation, string+int, string+float etc. was pretty
much copy-paste-implemented. Since we convert the int/float/etc to an
`std::string` already, we can take advantage of this and add a single
`std::string` implementation instead, that is called from the various
places needed. This was found while working on
https://gitlab.perlang.org/perlang/perlang/-/merge_requests/531, since
this change will make the fixes there much simpler to implement (fewer
places calling `malloc` that need to be replaced with `new[]` calls)

https://gitlab.perlang.org/perlang/perlang/-/merge_requests/536