(stdlib) Add `File.read_all_text()` method
...and use this method from the C# code for reading the Perlang program
being executed from disk. Doing it like this makes little "theoretical
sense" since it's much more cumbersome than just using the built-in C#
methods for this, *but* it means that we start using the new
C++/Perlang-based stdlib somewhere. In that sense, it's a small step
towards the effort of becoming self-hosted.

https://gitlab.perlang.org/perlang/perlang/-/merge_requests/527