(stdlib) Avoid unnecessary `strdup()` call in `stdlib_wrappers.cc`
The problem was that we didn't have any mechanism for "move" semantics
from the backing byte array in a `perlang::String` to the outside world
(since it was a plain, C-style `const char*` pointer). By utilzing the
goodness that `std::unique_ptr` brings to the table, we can just use a
`release()` call in the right place and enjoy the benefits of:

- One memory allocation (and deallocation) less
- One string copying less

It doesn't make much difference right now, since Perlang programs are
typically small, but what if we would use this method to read a 1
megabyte file? A 100 megabyte file? A 10 terabyte file? We should
*always* strive for writing efficient code, particularly when the cost
of doing so is reasonable and as long as it doesn't make the code messy.

One interesting consequence of this is that I had to change some
String-related methods no longer return a `const UTF8String` and
similar. I may have misunderstood the proper usage of `const` here. It's
probably correct to assume that "methods should seldom _return_ const
but often _receive_ a const parameter". In other words, when they don't
need mutation, make it possible to use `const` parameters, but don't
*force* immutability on the caller (by returning const) in cases where
some mutation might be necessary.

https://gitlab.perlang.org/perlang/perlang/-/merge_requests/538