(tests) Use `string[] arguments` parameter in `EvalHelper`
This is a preparational commit, preparing for adding a
`[CallerMemberName]` attribute to these methods. That attribute is hard
to combine with `params`-parameters, as described in
https://stackoverflow.com/questions/14355600/combining-callermembername-with-params.
Since we use the parameterized forms of these methods quite seldomly,
going with normal `string[]` arrays instead of `params string[]` is a
pretty small price to pay.

https://gitlab.perlang.org/perlang/perlang/-/merge_requests/541