(tests) Rewrite `cctest`-based tests with Catch2
We never really "deliberately" decided to use `cctest`; it happened to
come along more or less "accidentally" when we included the Google
`double-conversion` library in
https://gitlab.perlang.org/perlang/perlang/-/merge_requests/407. We had
a few basic tests there in place, but not a very exhaustive test
coverage of our C++-based `stdlib`.

Now, we are at the point where we would like to be able to write more
unit tests for the C++-based code. Catch2 seems like a useful library
for this. This change throws out all the remaining parts of the
`double-conversion` library (i.e. `cctest`), and adds Catch2 support. I
also added a simple Make target which runs the C++-based tests as part
of the CI runs. We couldn't really do it in
https://gitlab.perlang.org/perlang/perlang/-/merge_requests/407#note_1675
(because we still supported macOS at that point), but enabling it now
should be fine.

https://gitlab.perlang.org/perlang/perlang/-/merge_requests/543