(stdlib) Move `NativeFile` class to `stdlib`
This is a C# wrapper on top of the C++-based stdlib, so it might make
sense here. OTOH, the _wrapper_ is actually placed in `perlang_cli`. The
reason I started thinking about this is because I am starting to add
another C# wrapper for unmanaged code, and that code needs to be
accessible from the `Perlang.Interpreter` project.