(stdlib) Return non-`const` strings
As elaborated on in
https://gitlab.perlang.org/perlang/perlang/-/merge_requests/538,
returning `const` strings can be a bit limiting. This is probably not a
good way to write idiomatic C++ code; I'll readily admit that I haven't
done much C++ programming in the last 20 years.

This change adjusts all return types to use non-`const` strings instead.
The caller can then cast the `unique_ptr<String>` etc. instances to a
`const` value as needed, but the other way around is much more awkward.

https://gitlab.perlang.org/perlang/perlang/-/merge_requests/544