(stdlib) Add basic `StringBuilder` class
...and use it everywhere instead of the .NET class.

This can be used to construct Perlang strings with dynamic content. The
`StringBuilder` class uses a byte buffer behind the scenes, allowing the
string to be mutated until the user is "finished" with it. Once that
happens, an (immutable) string is produced which is then carried along
in the program.

https://gitlab.perlang.org/perlang/perlang/-/merge_requests/548