(stdlib) Support dynamically resizing `StringBuilder` buffer
It would previously use a fixed buffer of 1024 bytes; trying to use more
than this would cause exceptions to be thrown. Growing the buffer using
`realloc()` would have been possible, but the problem is that this would
have required the memory to have been allocated with `malloc()` in the
first place (i.e. not `new[]`).

Because of this, and because of suggestions seen on Stack Overflow, I
rewrote the code to use `std::vector` instead, which have a built-in
`resize()` operation. Also added a C++-based unit test to validate the
new functionality.

https://gitlab.perlang.org/perlang/perlang/-/merge_requests/550