(stdlib) Fix buffer overrun in dynamic resizing of `StringBuilder`
It has been said before, and it needs to be said again and again: C and
C++ are dangerous languages. :-) It's far too easy to shoot yourself in
the foot, and blow off the whole leg.

The previous algorithm for increasing the capacity *worked*, but only if
the string being added was small enough to fit into `capacity * 2`. The
new, fixed version should work regardless of size of the string being
added.

I'll also try to fix a CI job to prevent things like this getting merged
in the future, but in a separate commit.

https://gitlab.perlang.org/perlang/perlang/-/merge_requests/551