(stdlib) Fix memory leak in tests
This may seem minor and petty, but it causes a real problem now that we
start running the stdlib tests through Valgrind too. In other words, we
will do our very best to prevent *any* form of memory problems from
slipping into this code base. That's the theory, at least. :-)

https://gitlab.perlang.org/perlang/perlang/-/merge_requests/558