(ci) Run all GitHub Actions on `ubuntu-22.04` image
This is relevant because of
https://gitlab.perlang.org/perlang/perlang/-/commit/de8dbe3ace6adf1547ffcfb62c9600407728672a#note_3573,
and the build issues we are currently experiencing on gcc 13. I tried
installing gcc 13 locally, but the problem is that it's not yet
available on Debian stable (and installing it from Debian testing seems
messy, removing certain packages which are critical to me).

Let's postpone "properly" supporting gcc 13 until the package is easily
installable on Debian stable (and at that time, we probably *have to*
fix it since I'll otherwise be unable to build the C++ libs locally on
my machine :)