(perlang_cli) Move `TokenType` definition from C# to Perlang
This may seem like a small step, but in reality it's egregorious. It
takes the CppSharp (https://github.com/mono/CppSharp) project into use,
and uses it to add a new `auto-generated-bindings` Make target. When
this is executed, we *generate C# bindings* for `perlang_cli.cc` - which
is in turn generated from Perlang code. In other words, we have managed
to build a bridge from Perlang to C#!

The end result of this is that we can start migrating types from being
defined in C#, into being defined in Perlang, while still being able to
use them from C#. This is incredibly important to be able to avoid a
"big-bang"-style of change, where we convert the whole compiler in one,
massive commit. Instead, we'll be able to take it little by little.

Moving an enum like `TokenType` was a simple first step. Rewriting
C#-based logic into Perlang is likely to be more challenging.

https://gitlab.perlang.org/perlang/perlang/-/merge_requests/571