(interpreter) Rewrite `Func` to use `IBindingRetriever` interface
...instead of a `Func<Expr, Binding>`. This can be considered a design
choice or "style preference", but I believe it is more than that. The
`Func<Expr, Binding>` makes the different parts of the system very
losely coupled, which makes it harder than desirable to use "Go To
Definition"-style functionality in the editor. In other words, using an
interface here makes things slightly more "strongly typed" which I
believe is beneficial.

https://gitlab.perlang.org/perlang/perlang/-/merge_requests/572