(compiler) Add `class` support
At last, after almost 5 years of development, it's finally happening.
Classes are making their way (back) into the Perlang language, now that
we both:

- need them, and
- have a good way to represent them in the underlying compiled output

The latter part is kind of obvious. Since we are compiling Perlang to
C++, the natural representation in the compiled form is to match Perlang
classes 1-to-1 with C++ classes. This also makes a lot of sense in our
current interop approach; by exposing Perlang classes as C++ classes, we
should be able to use them from C# reasonably seamlessly, or that's the
idea at least.

https://gitlab.perlang.org/perlang/perlang/-/merge_requests/573