(compiler) Add `#include` statements to generated C++ headers
These were technically not "required" to be able to use the header, if
the consuming code would include the system-header before including the
generated C++ header file. Still, it would make it harder to consume the
header file on its own, so I think this is the right thing to do.